requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
h11==0.14.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
outcome==1.3.0.post0
pymongo==4.11.2
PySocks==1.7.1
python-dotenv==1.0.1
regex==2024.11.6
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
